+++ src/sha2/sha512/mod.rs
@@ -434,6 +434,19 @@ mod tests {
     }
 
     #[test]
+    fn test_verify() {
+        let digest = new().update("data").digest();
+
+        assert_eq!(new().update("data").verify(digest), true);
+        assert_eq!(new().update(b"data").verify(digest), true);
+        assert_eq!(new().verify(digest), false);
+
+        assert_eq!(new().update("data").finalize().verify(digest), true);
+        assert_eq!(new().update(b"data").finalize().verify(digest), true);
+        assert_eq!(new().finalize().verify(digest), false);
+    }
+
+    #[test]
     fn test_hello_world() {
         let digest = new().update("Hello World").digest();
         assert_eq!(digest.to_hex_lowercase(), "2c74fd17edafd80e8447b0d46741ee243b7eb74dd2149a0ab1b9246fb30382f27e853d8585719e0e67cbda0daa8f51671064615d645ae27acb15bfb1447f459b");